修改hashmap解决leetcode 136题代码，java集合类可以对数字自动装箱，不必进行强制类型转换

+++ practice/leetcode/101-200/136.single-number-hashmap-solution.ac.java
@@ -45,10 +45,10 @@ class Solution {
         int len = nums.length;
 
         for (int i = 0; i < len; i++) {
-            if (map.get((Integer)nums[i]) == null) {
-                map.put((Integer)nums[i], 1);
+            if (map.get(nums[i]) == null) {
+                map.put(nums[i], 1);
             } else {
-                map.remove((Integer)nums[i]);
+                map.remove(nums[i]);
             }
         }
